covertdtls: restore the widget wasm build, and add CI that would have caught it
The browser widget cannot be built from source on main:

    $ GOOS=js GOARCH=wasm go build ./cmd
    common/covertdtls/covertdtls.go:83:5: s.SetSRTPProtectionProfiles undefined
    common/covertdtls/covertdtls.go:84:5: s.SetDTLSClientHelloMessageHook undefined
    exit 1

covertdtls is imported by clientcore/producer.go and clientcore/settings.go —
core widget code, not a side path — and had no build tags, so it compiled for
js/wasm too. The pion SettingEngine hooks it installs don't exist there: in a
browser the DTLS handshake belongs to the browser's own WebRTC stack, so there is
no ClientHello for us to shape.

Two things kept this invisible. No CI job builds the wasm widget, and
ui/public/widget.wasm is a committed prebuilt binary, so the source path was
never exercised and every check stayed green.

Fix: split Apply into apply.go (!wasm, unchanged) and apply_wasm.go (wasm,
no-op). Only Apply touches the unavailable API — Config, ParseModeString and
Enabled are platform-independent and stay shared, so this is a narrow split
rather than tagging out the package.

An implication worth stating outright, because it is easy to miss and affects
what we believe about our own coverage: the package doc explains covertdtls
exists because the default pion/dtls fingerprint was being DPI-filtered in Russia
from 2026-03-30 (net4people/bbs#603). That evasion has never applied to browser
widgets, and cannot — browsers control DTLS. A widget config setting
randomize/mimic is silently ineffective. Fingerprint control requires a native
producer. The wasm stub returns nil rather than an error so a shared config
carrying a covertdtls mode doesn't fail widget startup over a capability the
platform can't provide.

Also adds .github/workflows/build-widget-wasm.yml, which compiles ./cmd for
js/wasm on PRs and pushes touching clientcore/common/cmd/netstate. Build only: a
go vet step under wasm tags currently reports four pre-existing clientcore
findings that native vet never type-checks, so including it would land the job
red on arrival. Noted in the workflow for follow-up.

Verified: js/wasm build of ./cmd exits 0; native go build ./... and the common
tests are unaffected.
